nfsd41: Modify server sessionid_hashtble
The sessionid will need to be resolved for every 4.1 COMPOUND RPC that has
an OP_SEQUENCE, and is therefore in the critical performance path.

We want a hash table with very few collisions (less than 2 on 80% of buckets).

How many sessionids do we expect per clientid?
- a client may only create a couple of sessions per server socket
- pNFS client will have a session per DS, and one for MDS.

Increase the sessionid_hashtable size to 1031 entries to start with.
We should investigate dynamically resizing the hash table size.

Use the crypto crc32 hash function to hash the sessionid.

Note: This code needs to be tested for hash distribution, and we could
investigate using a round-robin index scheme with current_session counter
as index.

Signed-off-by: root <root@sager-fedora.(none)>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>