nfsd41: Expand current_sid to include the current slot
Server upcall architecture can result in a partially processed COMPOUND plac
back into the receive queue to wait for upcall completion.
Add state to the slot structure to notice that an OP_SEQUENCE is beening
processed.

Reference count the nfs41_session structure. Hold a reference to the session
struct during COMPOUND processing.

Expand the current_sid in nfsd4_proc_compound to a current_session structure
Set the slot state and dereference the session at the end of COMPOUND
processing.

Signed-off-by: root <root@sager-fedora.(none)>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>