nfsd41: cleanup sessionid definition for client and server
sessionid_t was defined both in include/linux/nfsd/state.h and in
include/linux/nfs41_sessions.h. This patch moves the common definition
to nfs4.h renamed nfs41_sessionid as well as defining nfsd_sessionid_t
in nfsd/state.h as the formatted version of the sessionid as
used by the nfsd.

Signed-off-by: Andy Adamson <andros@umich.edu>
Signed-off-by: root <root@sager-fedora.(none)>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>