nfs41: Rearrange the enc_xdr functions in order to share them between NFSv4.0 and NFSv4.1.
This patch splits the xdr encoderoutines into routines that set up the
header and routines that fill in the ops. Hence, NFSv4.0 can set up a
compound with the minorversion set to 0 and then call the op filling
routine, while the equivalent NFSv4.1 can set the minorversion to 1
and insert in the SEQUENCE op before calling the op filling routine.

op before caqlling the op

Signed-off-by: Benny Halevy <bhalevy@panasas.com>