pnfs: nfs4_async_handle_error for data server
It does not make sense for the pNFS filelayout data servers which are not
mounted, and use the MDS struct nfs_server for all bookkeeping.

The filelayout data servers use struct nfs_client to reference the
struct nfs4_session, and have no associated struct nfs_server.

For data server nfs4_async_handle_error calls the nfs_server parameter will
be the MDS nfs_server structure. This is OK because the server struct is
only used on NFS_DELAY errors which are not allowed (yet!) on data server
communication.

This interface will be used in subsequent patches that implement the
filelayout I/O.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>