pnfs: don't call unimplemented operations
Remove unimplemented pnfs operations from pnfs_v41_clientops so that
the nfslayoutdriver works. Will add opeations as implemented.

The 2.6.24 sessions patch included unimplemented stubs of pnfs functions.
This results in an nfslayoutdriver that compiles, but segfaults if run.
This patch keeps pnfs struct nfs_rpc_ops, but reverts the pnfs functions
to the non-pnfs nfs functions, which allows the nfslayoutdriver to run, albeit
not as a pnfs module. As pnfs functions get implemented, they can replace
the nfs counterpart, and the nfslayoutdriver can be incrementally tested as
functionality gets ported.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>