pnfs: getdeviceinfo
The getdeviceinfo call is made by the client when a layoutget operation
returns an unknown device id.
Use the filehandle of the layoutget call as the current filehandle
for the getdeviceinfo call, instead of the filehandle of the pNFS mount point.

The 2.6.18.3 getdeviceinfo call used the root filehandle of the mount point
stored in struct nfs_server->fh. The 2.6.24 nfs client no longer saves the
root filehandle. Instead, a local variable is used.

There is no reason to use the MDS mountpoint filehandle for getdeviceinfo.
Instead of reverting the 2.6.24 code to save the filehandle of the pNFS
mount point, this patch uses the filehandle of the layoutget call that
initiated the getdeviceinfo call.

Change the interface of nfs4filelayoutdev.c:get_device_info, replacing the
struct filelayout_mount_point pointer to a struct inode pointer.
The inode pointer is used to get the filehandle for the getdeviceinfo call.

Declare a macro to dereference the struct filelayout_mount_type given an
inode.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>