pnfs: remove PNFS_USE_DS flag
In 2.6.18.3, there are several read I/O paths through the nfs client code.
The PNFS_USE_DS flag was set to have an easy way to check if
I/O was to be performed to the storage device.

In 2.6.24, there is only one read I/O path through the client code. The
interface to the nfs page cache has been re-written. It is now shared between
the read and write code paths, and moved entirely to fs/nfs/pagelist.c.

There is no need for the PNFS_USE_DS flag.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>