pnfs: allocate layout driver data in one piece with pnfs_layout_type
layout driver data comes right after the pnfs_layout_type fields
and can be accessed with the PNFS_LD_DATA accessor function.

The patch replaces the separate allocations of pnfslayout and
pnfslayout->layoutid with a single allocation of the unified
structure and replaces all references to pnfslayout->layoutid to
PNFS_LD_DATA(pnfslayout).

Signed-off-by: Benny Halevy <bhalevy@panasas.com>