pnfs: client layout cache: {get_lock,put_unlock}_current_layout
get_lock_current_layout locks and takes a reference count on pnfs_layout_type
put_unlock_current_layout reverses that.

get_alloc_layout allocates the layout and also returns the layout locked.

The higher-level callers (e.g. pnfs_update_layout) may (and should) release the
lock if they are going to block or run for an extended period of time.
The reference count on the layout guarantees that it will be kept around.
When done, the lock must be taken again via spinlock(&nfsi->lo_lock)
and put_unlock_current_layout must be called to release the reference
and unlock the layout.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>