pnfs: initialize pageio for read
The page cache can already have data in the readahead range. The nfs_readpages
pagelist will 'skip' these pages.  In this case using the number of pages in
the pagelist to calculate the layout range will be incorrect.

Use the first and last page in the pagelist to calculate the layout request
range.

The pageio coalesce code will create requests to fill the holes of data not
present in the page cache.

So! We could delay setting the rsize (and whether to use pNFS or RPC to MDS)
until the coalesce code, sending requests below the threshold to the MDS....

This code uses the whole readahead range for the threshold decision, as is
done in 2.6.18.3.

This patch calculates the readahead range, tries to get a layout, and if
successful, sets the coalesce size to ds_rsize.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>