pnfs: client filelayout_commit
If calling commit to the MDS, return 1 to pnfs_try_to_commit. If calling
commit to DS, set the correct DS address in the call to nfs_initiate_commit.

The commit code was passing in a strut list_head of pages as an argument,
but this is an empty list, its contents have already been moved to data->pages.

Try to make the pnfs_try_to_commit path more closely resemble the
pnfs_try_to_write path.

NOTE: The filelayout_commit() code gathers pages per DS  - I left this code
alone, but it should be reviewed because I think it is unnecessary because
the pages have already been gathered per DS in the write path.

Signed-off-by: Fred Isaman<iisaman@umich.edu>
Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>