nfsd41: create_session operation
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

pnfsd: fix compile error with gcc 3.4.4

From: Marc Eshel <eshel@almaden.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

nfsd41: nfsd4_create_session should return the seqid in the request

According to the NFSv4.1 specification, a CREATE_SESSION csr_sequence
must be equal to the csa_sequence in the request.  When we receive
a CREATE_SESSION for an unconfirmed clientid, we verify that the
csa_sequence is the same as the one we sent in our EXCHANGE_ID
reply.  Same as before this patch.

If the CREATE_SESSION request finds a confirmed clientid, and the
csa_sequence is equal to our confirmed sequence + 1, we then
increment our confirmed sequence and process the request.

Signed-off-by: Ricardo Labiaga <ricardo.labiaga@netapp.com>