pnfsd: Fix file layout layoutget export op for d13
Complete transformation of server handling of layoutget to be inline
with draft 13.  The main change is that the layout now only contains
a single deviceid, not an array of deviceids.  There is also
no device index list in the layout.

IMPORTANT: All file systems supporting the file layout type will need to update
their private layoutget export operation.

Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>