pnfsd: Implement getdevlist maxcount checking.
Implement the gdla_maxcount getdevicelist4args parameter.  Return
TOOSMALL if not even a single device fits in maxcount bytes.
This parameter is passed to the underlying file system to
determine how many devices it can return in a single getdevicelist
operation.

With file layout, lots of devices are sent under a single device id,
so the client will need to send a relatively large value of maxcount.

Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>