pnfs remove pnfs_writeback_done_update()
The function was called in the NON-RPC path in the 2.6.18.3 pNFS implementation.
Now, it is only called in the async write path to a file layout data server,
which will never require a call to pnfs_need_layoutcommit.

Remove pnfs_writeback_done_update and call pnfs_update_last_write()
from pnfs4_write_done.

Signed-off-by: Andy Adamson<andros@umich.edu>