pnfsd: get a reference on nfs4_file when cloning nfs4_layoutrecall
currently, the reference count is taken at hash_layoutrecall which
is the wrong time.  When the callback path is down the pending
nfs4_layoutrecall structure that was previously cloned is released
and since it points to the nfs4_file without a corresponding reference count
the nfs4_file strucutere is being released prematurely.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>
Tested-by: Marc Eshel <eshel@almaden.ibm.com>