pnfsd: Cleanup nfs4_pnfs_get_layout
Reorg nfs4_pnfs_get_layout to prevent unecessary
cleanup.

We already have a function to alloc the
layout, merging alloc and init in the same function
is unecessary.
Rename alloc_init_layout to init_layout.

If the layout is allocated before the export op
layout_get is called, it won't have the opportunity
to fail afterwards.
Therefore, by only initializing the layout after
layout_get, init_layout cannot fail, so it is
unecessary to free and return the layout upon failure.

Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>