nfs41: destroy_session operation
Signed-off-by: Ricardo Labiaga <ricardo.labiaga@netapp.com>
Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

nfs41: Add missing call to destroy session in nfs4_get_sb.

nfs4_get_sb will create a new session indirectly when creating a new
server structure.  It may later decide not to use it.  Make sure it
destroys the session in such case.

Need to investigate if we should move the session destroying into
nfs4_free_server.  I don't do it at this time since it is called from
many other places that do not currently create a session.  They should,
and the fix will be submited in a separate patch.

Signed-off-by: Ricardo Labiaga <ricardo.labiaga@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>