pnfs: xdr decode layoutget response logr_layout array count as per draft-19
If the logr_layout array count is zero, drop the reply as invalid.
If the coutn is greater than 1, decode only the first record at the moment
and drop the rest (and print a debug message).

FIXME: the whole array should be decoded and passed up to the pnfs client.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>