pnfs: fix pnfs4_*_done routines
The pnfs4_*_done routines are ignoring the possibility that the code
fell back to nfs.  This patch chops out layoutcommit setup that is done
there (incorrectly in the fallback case) which is already done by
pnfs_writepages anyway.

TO CHECK: pnfs4_read_done does not do a nfs_invalidate_atime that
nfs4_read_done does.  Does it need to be done in the fallback case?
In that case, do we need to add a flag to data->pnfsflags?

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>