pnfs: Fix commit through Data Server
The nfs_write_data structure is cloned for every Data Server before issuing
the commit call.  The res.verf should point to the structure's verfifier,
since this is where it will be stored (refer to nfs_commit_rpcsetup, where
it is first initialized).

This patch fixes a bug that had it pointing to the old structure verifier
which is eventually be freed.  Without this fix, commit through the Data
Server does not work, since the commit verifier in the reply is stored
in the wrong place.

Signed-off-by: Ricardo Labiaga <Ricardo.Labiaga@netapp.com>