pnfsd: Update getdeviceinfo for draft-19
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

pnfsd: encode empty getdeviceinfo notify bitmap rather than zeroed

If no notifications are given represent gdir_notification as an empty
bitmap array rather than one consisting of a single zeroed entry.

Thanks to Dean Hildebrand for suggesting this optimization
and to Peter Staubach for convincing that it's worth it.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>

pnfsd: do not depend on the current file handle in getdeviceinfo

starting draft-19 the client does not set the current file handle
in GETDEVICEINFO.  Since the deviceid namespace is global, the server
should determine the filesystem from the device ID

Signed-off-by: Marc Eshel <eshel@almaden.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>