pnfs: Move getdeviceinfo to draft-19 on client
When the layout driver calls get_device_info, it needs
to make sure it sets the id and type in the device
argument.

Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

[SQUASHME] pnfs: encode empty getdeviceinfo notify bitmap rather than zeroed

If no notifications are requested represent gdia_notify_types as an empty
bitmap array rather than one consisting of a single zeroed entry.

Thanks to Dean Hildebrand for suggesting this optimization
and to Peter Staubach for convincing that it's worth it.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>

pnfs: do not send putfh for getdeviceinfo

Starting draft-19 putfh is not needed for GETDEVICEINFO

Signed-off-by: Benny Halevy <bhalevy@panasas.com>