pnfs: remove pnfs_writeback_done_update()
The function was called in the NON-RPC path in the 2.6.18.3 pNFS implementation.
Now, it is only called in the async write path to a file layout data server,
which will never require a call to pnfs_need_layoutcommit.

Remove pnfs_writeback_done_update and call pnfs_update_last_write()
from pnfs4_write_done.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

pnfsd: Indicate layout needs to be commited if data is written

Currently we only invoke pnfs_need_layoutcommit() in the commit path.
This is incorrect, since the server may return stable writes, and our
client will never issue a commit.  In this case the file size will not
get updated in the MDS.  This patch fixes this problem.

Signed-off-by: Ricardo Labiaga <Ricardo.Labiaga@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>