pnfs: client rpc commit done
Implement the rpc call done callback for pNFS commit to DS or MDS.

With nfs_initiate_commit() there is no need for pnfs4_proc_commit_setup().

Signed-off-by: Andy Adamson<andros@umich.edu>

pnfs: no need to mark for layout commit at pnfs commit done time

now that need_layoutcommit is done on the write_done path for
writes to the DS there's no need to do that on the commit path.
This is also totally wrong for non-file layout types.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>