pnfs: file layout commit use wb_private
Remove struct nfs_page wb_devip and wb_devport fields. Refactor
filelayout_commit to use the struct nfs4_pnfs_dserver in the nfs_page
wb_private field.

Signed-off-by: Andy Adamson<andros@umich.edu>

pnfs: Fix commit through Data Server

The nfs_write_data structure is cloned for every Data Server before issuing
the commit call.  The res.verf should point to the structure's verfifier,
since this is where it will be stored (refer to nfs_commit_rpcsetup, where
it is first initialized).

This patch fixes a bug that had it pointing to the old structure verifier
which is eventually be freed.  Without this fix, commit through the Data
Server does not work, since the commit verifier in the reply is stored
in the wrong place.

Signed-off-by: Ricardo Labiaga <Ricardo.Labiaga@netapp.com>

pnfs: fix filelayout_clone of fattr

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>