pnfsd: layoutcommit set size fix
nfsd4_layoutcommit() should only reset the size of the file if
the client modified the file size.  The client indicates size change
by setting the no_newoffset field in the LAYOUT_COMMIT message.

Note that the 'no_' prefix stands for "new offset", not for negative.
If no_newoffset is false while XDR decoding (no file change), we set
lcp->lc_newoffset to zero to indicate this.

Signed-off-by: Ricardo Labiaga <Ricardo.Labiaga@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>