pnfs: decode_device error path bug fix
INIT_HLIST_NODE sets the next and prev pointers to NULL. hlist_del_rcu
references these pointers, and causes a crash in destroy_device when
the hlist is empty as is the case int the out_err_free path of decode_device.

Remove INIT_HLIST_NODE which is not needed due to kzalloc. Replace
hlist_del_rcu with hlist_del_init which checks for the NULL values.
Clean up comment in nfs4_pnfs_dev_item declaration.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>