panlayout: {read,write}_pagelist
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

panlayout: remove PNFS_ISSYNC

PNFS_ISSYNC is no longer used, so remove references to it

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

panlayout: repair break done by change in call API.

This restores behavior to what it was.  Note in particular that
proper cleanup via done routines is still not done.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

panlayout: implement 64-bit atomic ops for non-64-bit architectures.

delta_space_used was defined as atomic64_t which is not avaiable on
non-64 bit architectures.  Define it using a spin_lock and a signed 64-bit
counter where atomic64_t is not avaiable.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>