pnfs: read path set ds_size
Use the pNFS ds_rsize for read pageio.

When using the nfs page cache for pNFS I/O, it is necessary to setupe pages
using the rsize of the storage device instead of the rsize of the MDS.

In 2.6.18.3, there are several read I/O paths through the client code and
the rsize was used in each of these paths to setup the nfs page cache. Since
several functions used the rsize, an rpc_ops->rsize function makes sense.

In 2.6.24, there is only one read I/O path through the client code. The
interface to the nfs page cache has been re-written. It is now shared between
the read and write code paths, and moved entirely to fs/nfs/pagelist.c.

As a result, there is now only one place to set the rsize.

This patch removes the rpc_ops->rsize function pointer and
sets the rsize to the pNFS ds_size when pNFS is being used.

pNFS is not being used if layoutget fails, or the count is below the
threshold.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

pnfs: fix compile problems if CONFIG_PNFS turned off - read.c

Mostly, #include of header files should not encosed in #ifdef,
since *.h files themselves have the appropriate #ifdefs, and they may want
to define no-op versions of some functions.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>