pnfsd: Simplify device export ops.
The current device export ops retrieve the device, encode it, and them free it,
which is quite cumbersome and requires extra data copies. We want to move towards
an interface like readdir, which has only a single vfs op,

This patch replaces get_deviceinfo, devaddr_free, and devaddr_encode with
get_device_info and get_device_iter. get_device_info will
encode the device info for the supplied device id onto the output stream.

The retrieval of the device from the underlying file system
now occurs in the encoding function. In proc we just
need to ensure getdevinfo is supported by the file system
and set the current file handle.

get_device_info takes a function pointer pnfs_encodedev_t as an argument.
This argument allows NFSD to supply a function to encode the device, instead
of having the file system do the encoding.  The file system takes the
pnfs_encodedev_t argument and uses it to encode its devices.

pnfs_encodedev_t is only useful for layout types which exist in the kernel,
i.e., only file layout for now, object and block will set this
function pointer to NULL.  In the future, when object or block
is in the kernel, they can use this function to move
device encoding into the common nfsd code.

get_device_iter is used as an iterator to retrieve all devices
from the underlying file system.
Once the iterator returns a device id, the get_device_info
export op can be used to retrieve and encode the information for
the device. Use the cookie and verifier as the starting
point of the iterator. Given a cookie and a verifier, this
operation returns the associated deviceid, and the next cookie and
verifier.  If there are no more device ids, eof is returned.

This patch requires file systems that support the file layout type and
want to use the NFS server export interface to update the data
structures they create (pnfs_filelayout_device, pnfs_filelayout_multipath,
pnfs_filelayout_devaddr). These data structures are then used by
filelayout_encode_devinfo to encode the device.

Note: with file layout, lots of devices are sent under a single device id,
so the client will need to send a relatively large value of maxcount.

Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>

pnfsd: Remove device enc/free export ops

Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

pnfs: fix compile problems if CONFIG_PNFS turned off - exportfs.h

Mostly, #include of header files should not encosed in #ifdef,
since *.h files themselves have the appropriate #ifdefs, and they may want
to define no-op versions of some functions.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>