pnfs: layout driver ops existence macro helpers
Macros to be used in stubs for pnfs functions defined as static inline
functions in fs/nfs/pnfs.h with a no-op implementation when CONFIG_PNFS
is not defined.

The motivations are to have the non-pnfs no-op definition so that the
invocation need not be surrounded in #ifdef and to short circuit the
function call if the layout driver does not define a particular method.

Extracted from Fred's "pnfs: pnfs_write_begin" patch.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>