pNFS: Merge filelayout_flush_one into filelayout_write_pagelist.
details: The flush_one layout driver interface function is a specialized
function only for the file layout.  It's purpose was to calculate
the correct data server for each write request so this calculation
can be avoided when the data is actually sent on the wire.  It
was decided that having such a specialized function (plus its
associated private state) was not worth
the effort.  In addition, O_DIRECT needs to redo this calculation
anyways.

Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>