[SQUASHME] nfs41: Correctly use new session in exception handling
We will need to handle several layers of session recovery depending on the
error. As a last resort, we will need to use the big hammer; destroy
an existing session and create a new one.

The current code creates a new session without destroying the old one.
Destroy the existing session and initialize a new session before calling
nfs41_recover_session_sync.

Signed-off-by: Andy Adamson<andros@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>